Add files via upload
A whole new approach, it creates plots of the X and Y center of mass, by looking at these we should be able to identify various patterns/events. I have also uploaded some of the plots, their names are their descriptions, so cascadeBackcrosses is cascade for about half the time, then backcrosses for half the time. The ballStill plots was just a ball sitting on a ladder, but we found that just moving around was enough to cause the ball to barely vibrate, but it showed up on the plot. ballStillWeight has a weight on it so it doesnt shake. crazy stuff, so much here to explore! Thanks Ian!